add enzyme, update frontend testing, organize files
