+++ src/components/tests/App.test.js
@@ -1,6 +1,6 @@
 import React from "react";
 import ReactDOM from "react-dom";
-import App from "./App";
+import App from "../../App";
 
 test("renders learn react link", () => {
   const div = document.createElement("div");
